% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{gf}
\alias{gf}
\title{String interpolation to build regression formulas from vectors with variable names.}
\usage{
gf(form, values = list(), enclos = parent.frame(), as.formula = TRUE,
  form.env = parent.frame(), collapse = " + ")
}
\arguments{
\item{form}{A formula or string of a formula. E.g. \code{y~x+{controls}} where \code{controls} is a character vector with variables that will be inserted.}

\item{values}{A named list containing the vectors of variable names. If empty the enclosing environment will be searched.}

\item{enclos}{the enclosing environment in which the vectors with variable names will be taken if not provided by values. By default the calling enviornment.}

\item{as.formula}{if \code{TRUE} (default) return a formula object, otherwise a string.}

\item{form.env}{The environment assigned to the returned formula. By default the calling environment.}

\item{collapse}{how shall variables in a vector be collapsed. Default \code{" # "}.}
}
\description{
Insert vectors with variable names into \{ \} placeholder in a formula and collapse vectors with +.
}
\details{
An adaption of the \code{glue} function specialized to build
regression formulas from variable names stored in vectors.
}
\examples{
# Assume we want to estimate a demand function
# using ivreg

contr = c("x1","x2","x3") # exogenous control variables
instr = c(contr, "z1","z2") # instruments

# Replace {contr} and {instr} in formula
gf(q ~ p + {contr} | {instr})

# You could also provide the formula as string
gf("q ~ p + {contr} | {instr}")

}
