% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{varnames_snippet}
\alias{varnames_snippet}
\title{Creates a simple snippet code to generate a string
vector of all column names of a data frame}
\usage{
varnames_snippet(x)
}
\arguments{
\item{x}{A data frame or other objects with \code{names}.}
}
\description{
The if the data frame has 3 columns named
"x", "y" and "name with space"
the function cats and on windows also copies to
your clipboard the
following code:
}
\details{
\code{c("x","y","`name with space`")}

This can be helpful to specify vectors of control
variables when using \code{\link{gf}}. If there are
many variables it is just easier to remove not-used variables
from the created code snippet than to write down
manually all used variables.
}
\examples{
df = data.frame(x=1,y=5,z=4,`a b`=4)
varnames_snippet(df)

}
