% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{gf}
\alias{gf}
\title{String interpolation of variables in regression formulas}
\usage{
gf(form, values = list(), enclos = parent.frame(), as.formula = TRUE,
  form.env = parent.frame(), collapse = " + ")
}
\arguments{
\item{form}{A formula or string of a formula. E.g. \code{y~x+{controls}} where \code{controls} is a character vector with variables that will be inserted.}

\item{values}{A named list containing the vectors of variable names. If empty the enclosing environment will be searched.}

\item{enclos}{the enclosing environment in which the vectors with variable names will be taken if not provided by values. By default the calling enviornment.}

\item{as.formula}{if \code{TRUE} (default) return a formula object, otherwise a string.}

\item{form.env}{The environment assigned to the returned formula. By default the calling environment.}

\item{collapse}{how shall variables in a vector be collapsed. Default \code{" # "}.}
}
\description{
Insert character vectors with variable names into a formula collapsing vectors with +. Similar to the \code{glue} function
from the package glue.
}
